+++ tests/src/test/resources/test-run-cron.php
@@ -5,8 +5,4 @@
  *
  * Please see LICENCE.md for complete licence text.
  */
-require_once __DIR__ . '/../include/functions.inc.php';
-require_once __DIR__ . '/../include/database.class.php';
-
-ob_start();
 include __DIR__ . '/../cronjobs/cron.php';